backend/requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.2
colorama==0.4.6
cryptography==46.0.5
Django==6.0.4
django-cors-headers==4.9.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.111.0
fastapi-cli==0.0.24
google-ai-generativelanguage==0.6.4
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.5.4
googleapis-common-protos==1.74.0
graphifyy==0.4.13
grpcio==1.80.0
grpcio-status==1.62.3
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
networkx==3.6.1
numba==0.65.0
numpy==2.4.2
opencv-python==4.13.0.92
openpyxl==3.1.5
orjson==3.11.8
packaging==26.0
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pefile==2024.8.26
pillow==12.1.1
platformdirs==4.9.6
pooch==1.9.0
proto-plus==1.27.2
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
PyMatting==1.1.15
pyparsing==3.3.2
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.26
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
rembg==2.0.75
requests==2.33.1
rich==15.0.0
rich-toolkit==0.19.7
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
starlette==0.37.2
tifffile==2026.4.11
tqdm==4.67.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-c-sharp==0.23.1
tree-sitter-cpp==0.23.4
tree-sitter-elixir==0.3.5
tree-sitter-go==0.25.0
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-julia==0.23.1
tree-sitter-kotlin==1.1.0
tree-sitter-lua==0.5.0
tree-sitter-objc==3.0.2
tree-sitter-php==0.24.1
tree-sitter-powershell==0.26.3
tree-sitter-python==0.25.0
tree-sitter-ruby==0.23.1
tree-sitter-rust==0.24.2
tree-sitter-scala==0.24.0
tree-sitter-swift==0.0.1
tree-sitter-typescript==0.23.2
tree-sitter-verilog==1.0.3
tree-sitter-zig==1.1.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.29.0
watchfiles==1.1.1
websockets==16.0
